05.06-compartilhando-os-jogos-recomendados

+++ src/main/kotlin/br/com/alura/alugames/modelo/Recomendavel.kt
@@ -1,7 +1,9 @@
 package br.com.alura.alugames.modelo
 
+import java.math.BigDecimal
+
 interface Recomendavel {
-    val media: Double
+    val media: BigDecimal
 
     fun recomendar(nota: Int)
 }
